Bandits package is ready for testing

+++ src/gossipLearning/models/bandits/BanditModel.java
@@ -4,6 +4,6 @@ import gossipLearning.interfaces.Model;
 
 public interface BanditModel extends Model {
   public double predict(int armIdx);
-  public int numberOfPlayes(int armIdx);
+  public long numberOfPlayes(int armIdx);
   public long numberOfAllPlayes();
 }
